pyproject.toml
[project]
name = "cv-final"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.127.0",
    "openai>=2.14.0",
    "opencv-python-headless>=4.12.0.88",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
